new feature: now DurationStatistic is creating an uniqueID for each interaction
Duration statistic is creating an uniqueID for each interaction with ID
defined in stattype. In case of stattype ID is not unique (for instance
connection socket id could be reused) uniqueID guarantees that each
transaction even having the same stattype ID is calculated separately.
This applies to stattypes defined with startwith/endwith tags only.